Discrete weights, new strategy
The model now trains and discretizes the weights during the validation step of training. Accuracy has increased to high 80s-90s, and as far as I know the final weights are unchanged during the discretization step of the finished model, before it checks the testing data. Model currently training 200 epochs